Fix: Project structure and build
