+++ src/components/elements/InfoItem/styles.ts
