+++ src/components/elements/PercentageText/PercentageText.tsx
