+++ src/components/elements/PercentageText/index.ts
