+++ src/components/elements/PercentageText/types.ts
