+++ src/components/elements/SectionItem/styles.ts
