+++ src/components/elements/SectionItem/types.ts
