+++ src/components/elements/Sparklines/types.ts
