+++ src/components/elements/index.ts
@@ -1,0 +1,7 @@
+import InfoItem from './InfoItem';
+import PercentageText from './PercentageText';
+import SectionItem from './SectionItem';
+import Sparklines from './Sparklines';
+import Table from './Table';
+
+export { InfoItem, PercentageText, SectionItem, Sparklines, Table };
