--- src/components/index.ts
@@ -1,5 +1,0 @@
-import PercentageText from './PercentageText';
-import Sparklines from './Sparklines';
-import Table from './Table';
-
-export { PercentageText, Sparklines, Table };
