+++ src/components/modules/Body/types.ts
@@ -1,0 +1,5 @@
+import type { AppTab } from '@/hooks/useTabs/types';
+
+export type BodyProps = {
+  activeTab: AppTab;
+};
