+++ src/components/modules/Header/types.ts
@@ -1,4 +1,4 @@
-import type { AppTab } from '../../hooks/useTabs/types';
+import type { AppTab } from '@/hooks/useTabs/types';
 
 export type HeaderProps = {
   // eslint-disable-next-line no-unused-vars
