+++ src/components/modules/Tabs/Coins/index.ts
