+++ src/components/modules/Tabs/Coins/styles.ts
