+++ src/components/modules/index.ts
@@ -1,0 +1,6 @@
+import Body from './Body';
+import Header from './Header';
+import CoinsTab from './Tabs/Coins';
+import ExchangesTab from './Tabs/Exchanges';
+
+export { Body, Header, CoinsTab, ExchangesTab };
