+++ src/components/templates/index.ts
@@ -1,0 +1,3 @@
+import MainTemplate from './MainTemplate/MainTemplate';
+
+export { MainTemplate };
