+++ src/hooks/index.ts
@@ -1,0 +1,6 @@
+import useCoinGeckoApi from './useCoinGeckoApi';
+import useCoins from './useCoins';
+import useGlobalData from './useGlobalData';
+import useTabs from './useTabs';
+
+export { useCoinGeckoApi, useCoins, useGlobalData, useTabs };
