+++ src/hooks/useGlobalData/index.ts
