--- src/pages/Main/components/Body/components/Tabs/Coins/hooks/index.ts
@@ -1,3 +1,0 @@
-import useCoins from './useCoins';
-
-export { useCoins };
