--- src/pages/Main/components/Body/components/index.ts
@@ -1,5 +1,0 @@
-import InfoItem from './InfoItem';
-import CoinsTab from './Tabs/Coins';
-import ExchangesTab from './Tabs/Exchanges';
-
-export { InfoItem, CoinsTab, ExchangesTab };
