--- src/pages/Main/components/Body/hooks/index.ts
@@ -1,3 +1,0 @@
-import useGlobalData from './useGlobalData';
-
-export { useGlobalData };
