--- src/pages/Main/components/Body/types.ts
@@ -1,5 +1,0 @@
-import type { AppTab } from '../../hooks/useTabs/types';
-
-export type BodyProps = {
-  activeTab: AppTab;
-};
