--- src/pages/Main/components/Header/components/index.ts
@@ -1,3 +1,0 @@
-import SectionItem from './SectionItem';
-
-export { SectionItem };
