--- src/pages/Main/components/index.ts
@@ -1,4 +1,0 @@
-import Body from './Body';
-import Header from './Header';
-
-export { Header, Body };
