--- src/pages/Main/hooks/index.ts
@@ -1,4 +1,0 @@
-import useCoinGeckoApi from './useCoinGeckoApi';
-import useTabs from './useTabs';
-
-export { useCoinGeckoApi, useTabs };
